Revert "ArchiveCommit - before reverting to prior"
This reverts commit 0088532285d1d2fa70157aa2a5602743dc6e4833.

+++ Assets/---------------Scripts------------/---------------UI---------------/LifeBar.cs
@@ -5,16 +5,17 @@ using UnityEngine.UI;
 
 public class LifeBar : MonoBehaviour
 {
-    // The worlds most simple life bar slider script O_O - a bit buggy right now.
+    // the worlds most simple life bar slider script O_O - a bit buggy right now.
     public Slider slider;
-    public void SetMaxLife(float life)
+    public void SetMaxLife(int life)
     {
         slider.maxValue = life;
         slider.value = life;
     }
 
-    public void SetLife(float life)
+    public void SetLife(int life)
     {
         slider.value = life;
     }
+
 }
